requirements.txt
# 浏览器自动化
playwright>=1.40.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# HTTP / LLM
httpx>=0.25.0
pyyaml>=6.0.0

# 图像（按需与 paddle 版本对齐）
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
